created barcodes API route and instead fetching data from that API to now render the page
