+++ package.json
@@ -12,7 +12,8 @@
     "next": "9.5.3",
     "react": "16.13.1",
     "react-dom": "16.13.1",
-    "sass": "^1.27.0"
+    "sass": "^1.27.0",
+    "swr": "^0.3.5"
   },
   "devDependencies": {
     "prettier": "2.1.2"
